requirements.txt
streamlit
PyPDF2
google-generativeai
langchain
langchain-google-genai
faiss-cpu
python-dotenv
langchain-community
